refactor: :recycle: Number input refactor. Forwardref included
