+++ libs/dana/src/components/index.ts
@@ -26,3 +26,4 @@ export * from './Card';
 export * from './Divider';
 export * from './Table';
 export * from './TextInput';
+export * from './NumberInput';
